requirements.txt
fastapi==0.119.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.24
asyncpg==0.30.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pytest-asyncio>=0.23.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
httpx==0.25.2
pytest-mock==3.12.0
email-validator>=2.0.0